examples/python-kafka-avro/requirements.txt
avro-python3
kafka-python
pandas